instead test parsing all DBCs
